runtime: reuse Go 1.26 Windows syscall wrappers
