+++ runtime/internal/lib/runtime/syscall_windows_llgo.go
@@ -60,80 +60,6 @@ func syscall_syscalln(fn, n uintptr, args ...uintptr) (r1, r2, err uintptr) {
 	return call.r1, call.r2, call.err
 }
 
-// The deprecated fixed-width entry points still appear in generated Windows
-// syscall code. nargs, rather than the wrapper width, controls how many
-// arguments are visible to the callee.
-
-//go:linkname syscall_Syscall syscall.Syscall
-//go:nosplit
-func syscall_Syscall(fn, n, a1, a2, a3 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3)
-}
-
-//go:linkname syscall_Syscall6 syscall.Syscall6
-//go:nosplit
-func syscall_Syscall6(fn, n, a1, a2, a3, a4, a5, a6 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3, a4, a5, a6)
-}
-
-//go:linkname syscall_Syscall9 syscall.Syscall9
-//go:nosplit
-func syscall_Syscall9(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9)
-}
-
-//go:linkname syscall_Syscall12 syscall.Syscall12
-//go:nosplit
-func syscall_Syscall12(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12)
-}
-
-//go:linkname syscall_Syscall15 syscall.Syscall15
-//go:nosplit
-func syscall_Syscall15(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15)
-}
-
-//go:linkname syscall_Syscall18 syscall.Syscall18
-//go:nosplit
-func syscall_Syscall18(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15, a16, a17, a18 uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, n, a1, a2, a3, a4, a5, a6, a7, a8, a9, a10, a11, a12, a13, a14, a15, a16, a17, a18)
-}
-
-//go:linkname syscall_SyscallN syscall.SyscallN
-//go:nosplit
-func syscall_SyscallN(fn uintptr, args ...uintptr) (r1, r2, err uintptr) {
-	return syscall_syscalln(fn, uintptr(len(args)), args...)
-}
-
-//go:linkname c_loadLibrary C.llgo_load_library
-func c_loadLibrary(filename *uint16, flags uint32, errno *uint32) uintptr
-
-//go:linkname c_getProcAddress C.llgo_get_proc_address
-func c_getProcAddress(handle uintptr, name *uint8, errno *uint32) uintptr
-
-//go:linkname syscall_loadlibrary syscall.loadlibrary
-func syscall_loadlibrary(filename *uint16) (handle, err uintptr) {
-	var errno uint32
-	handle = c_loadLibrary(filename, 0, &errno)
-	return handle, uintptr(errno)
-}
-
-//go:linkname syscall_loadsystemlibrary syscall.loadsystemlibrary
-func syscall_loadsystemlibrary(filename *uint16) (handle, err uintptr) {
-	const loadLibrarySearchSystem32 = 0x00000800
-	var errno uint32
-	handle = c_loadLibrary(filename, loadLibrarySearchSystem32, &errno)
-	return handle, uintptr(errno)
-}
-
-//go:linkname syscall_getprocaddress syscall.getprocaddress
-func syscall_getprocaddress(handle uintptr, name *uint8) (proc, err uintptr) {
-	var errno uint32
-	proc = c_getProcAddress(handle, name, &errno)
-	return proc, uintptr(errno)
-}
-
 // Windows callbacks need a separate C-to-Go ABI thunk and are implemented in
 // the callback/FFI stage of Windows support. Keep the unsupported behavior
 // explicit instead of returning a pointer that would fail nondeterministically.
